Выполнил 5 задание
