+++ src/film-detalis.js
@@ -1,7 +1,8 @@
-import createElement from '../src/create-element.js';
+import Component from '../src/component.js';
 
-export default class {
+export default class FilmDetalis extends Component {
   constructor(data) {
+    super();
     this._image = data.image;
     this._title = data.title;
     this._description = data.description;
@@ -11,12 +12,15 @@ export default class {
     this._genre = data.genre;
     this._comments = data.comments;
 
-    this._element = null;
-    this._onClick = null;
+    this._onCloseButtonClick = this._onCloseButtonClick.bind(this);
   }
 
   _getGenre() {
     return [...this._genre][Math.floor(Math.random() * 5)];
+  }
+
+  _onCloseButtonClick() {
+    this._onClick();
   }
 
   get template() {
@@ -185,27 +189,11 @@ export default class {
     </section>`;
   }
 
-  set onClick(func) {
-    if (typeof func === `function`) {
-      this._onClick = func;
-    }
-  }
-
   bind() {
-    this._element.querySelector(`.film-details__close-btn`).addEventListener(`click`, this._onClick.bind(this));
+    this._element.querySelector(`.film-details__close-btn`).addEventListener(`click`, this._onCloseButtonClick);
   }
 
-  get element() {
-    return this._element;
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this._element = null;
+  unbind() {
+    this._element.querySelector(`.film-details__close-btn`).removeEventListener(`click`, this._onCloseButtonClick);
   }
 }
